requirements.txt
fastapi
uvicorn[standard]
websockets
pillow
aiohttp
pywin32; sys_platform == "win32"
